MINOR: sample: Add sample fetches for enhanced observability for TLS ClientHello
Add new sample fetches to get the ciphers, supported groups, key shares and signature algorithms
that the client supports during a TLS handshake as part of the contents of a TLS ClientHello.
Currently we can get the following contents of the ClientHello message: SNI(req_ssl_sni) and
TLS protocol version(req_ssl_ver). The following new sample fetches will be added to get the
following contents of the ClientHello message exchanged during the TLS handshake (supported by
the client):
- req.ssl_cipherlist: Returns the binary form of the list of symmetric cipher options
- req.ssl_sigalgs: Returns the binary form of the list of signature algorithms
- req.ssl_keyshare_groups: Return the binary format of the list of key share group
- req.ssl_supported_groups: Returns the binary form of the list of supported groups used in key exchange

This added functionality would allow routing with fine granularity pending the capabilities the client
indicates in the ClientHello. For example, this would allow the ability to enable TLS passthrough or
TLS termination based on the supported groups detected in the ClientHello message. Another usage is to
take client key shares into consideration when deciding which of the client supported groups should be
used for groups considered to have 'equal security level' as well as enabling fine grain selection of
certificate types(beyond the RSA vs ECC distinction). All of that is relevant in the context of rapidly
upcoming PQC operation modes.

Fixes: #2532